+++ spec/timeout_spec.rb
@@ -1,15 +1,26 @@
 require 'awesome_bot'
 
 describe AwesomeBot do
-  describe "check" do
-    context "given a timeout link" do
-      options = {'timeout'=>1}
-      r = AwesomeBot::check 'http://www.cmr.osu.edu/browse/datasets', options
+  describe "timeout" do
+    timeoutlink = 'http://www.cmr.osu.edu/browse/datasets'
+    options = {'timeout'=>1}
+
+    context "given a timeout link and setting timeout to 1s" do
+      r = AwesomeBot::check timeoutlink, options
       s = r.statuses_issues[0]
       expected = 'execution expired'
       value = s['error'].to_s
       it "has a timeout" do
         expect(value).to include(expected)
+      end
+    end
+
+    context "given a timeout link and setting timeout to 1s with --allow-timeout" do
+      r = AwesomeBot::check timeoutlink, options
+      expected = 0
+      value = r.statuses_issues(false, true).count
+      it "has no issues" do
+        expect(value).to eql(expected)
       end
     end
   end